Conflicts:
	ceph.spec.in

+++ src/librbd/internal.h
@@ -165,7 +165,7 @@ namespace librbd {
   void init_rbd_header(struct rbd_obj_header_ondisk& ondisk,
 		       uint64_t size, int order, uint64_t bid);
 
-  int64_t read_iterate(ImageCtx *ictx, uint64_t off, size_t len,
+  int64_t read_iterate(ImageCtx *ictx, uint64_t off, uint64_t len,
 		       int (*cb)(uint64_t, size_t, const char *, void *),
 		       void *arg);
   int diff_iterate(ImageCtx *ictx, const char *fromsnapname,
